+++ app/Http/Controllers/Admin/LocationController.php
@@ -7,25 +7,30 @@ namespace App\Http\Controllers\Admin;
 use App\Enums\OperationsEnum;
 use App\Http\Controllers\Controller;
 use App\Models\Location;
+use App\Models\User;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Validation\Rule;
 use Inertia\Inertia;
 use Inertia\Response;
 
 final class LocationController extends Controller
 {
+    public function __construct(private User $user) {}
     /**
      * Display a listing of the resource.
      */
     public function index(Request $request): Response
     {
+        Gate::authorize('view_any_location', $this->user);
+
         $locations = Location::query()
             ->when(
                 $request->search,
-                fn ($q) => $q->where('city', 'like', '%'.$request->search.'%')
-                    ->orWhere('state', 'like', '%'.$request->search.'%')
-                    ->orWhere('country', 'like', '%'.$request->search.'%')
+                fn($q) => $q->where('city', 'like', '%' . $request->search . '%')
+                    ->orWhere('state', 'like', '%' . $request->search . '%')
+                    ->orWhere('country', 'like', '%' . $request->search . '%')
             )
             ->paginate($request->perPage ?? 10)
             ->withQueryString();
@@ -41,6 +46,8 @@ final class LocationController extends Controller
      */
     public function create(): Response
     {
+        Gate::authorize('create_location', $this->user);
+
         $operation = OperationsEnum::Create;
 
         return Inertia::render('admin/locations/create-or-edit-location', [
@@ -54,13 +61,15 @@ final class LocationController extends Controller
      */
     public function store(Request $request): RedirectResponse
     {
+        Gate::authorize('create_location', $this->user);
+
         $data = $request->validate([
             'city' => ['required', 'string', 'max:255'],
             'state' => ['nullable', 'string', 'max:255'],
             'country' => ['required', 'string', 'max:255'],
 
             Rule::unique('locations')->where(
-                fn ($query) => $query->where('city', $request->city)
+                fn($query) => $query->where('city', $request->city)
                     ->where('state', $request->state)
             ),
         ], [
@@ -85,6 +94,8 @@ final class LocationController extends Controller
      */
     public function edit(Location $location): Response
     {
+        Gate::authorize('update_location', $this->user);
+
         $operation = OperationsEnum::Edit;
 
         return Inertia::render('admin/locations/create-or-edit-location', [
@@ -99,6 +110,8 @@ final class LocationController extends Controller
      */
     public function update(Request $request, Location $location): RedirectResponse
     {
+        Gate::authorize('update_location', $this->user);
+
         $data = $request->validate([
             'state' => ['nullable', 'string', 'max:255'],
             'country' => ['required', 'string', 'max:20'],
@@ -107,7 +120,7 @@ final class LocationController extends Controller
                 'string',
                 'max:255',
                 Rule::unique('locations')
-                    ->where(fn ($query) => $query->where('state', $request->state))
+                    ->where(fn($query) => $query->where('state', $request->state))
                     ->ignore($location->id),
             ],
         ], [
@@ -124,6 +137,8 @@ final class LocationController extends Controller
      */
     public function destroy(Location $location): RedirectResponse
     {
+        Gate::authorize('delete_location', $this->user);
+
         $location->delete();
 
         return to_route('admin.locations.index')->with('success', 'Location record deleted successfully.');
